k8s: Moved CRD code to v2 pkg
 - Moved common code to utils.go
